Fix wrong answer bug
When you choose the wrong flag, your score should
decrease with 1 point. It would do that, but
after a newly generated question, even if you choose
the right answer, you would lose a point and get
back to 0 points

I created a div with the id 'game-header' that
contains both the question and the country so
the options div could be empty. After that,
in the script.js file, I would empty that div
after the correct flag is chosen and reload it
with another set of flags